Add .editorconfig
I've tested that this file matches current coding styles in the repo so that no warnings are introduced.
The reason I noticed the lack of this file is that the csproj had *one* tab in it, and spaces for indentation everywhere else. This is a telltale sign that the editor doesn't know whether to use tabs when you manually edit the project file. By adding this .editorconfig file, inconsistencies such as this and many others are avoided.

I also add a `Directory.Build.props` file so that any code style deviations will appear as warnings in the build. Then I modify the GitHub Actions workflow to treat warnings as errors to break the build so that any pull request that degrades your consistent coding style will fail validation.