Do not rebuild when packing in GitHub Actions
`dotnet pack` will build everything again, modulo whatever msbuild's 'incremental build' wants to skip, which isn't great.
Adding `--no-build` to the `dotnet pack` command is better, but it is a bit fragile and can cause build failures if your build authoring ever trips over invoking the build target.
The best solution to this is usually to build and pack at the same time, since that's reliable and fast.